routing/pathfinding/requirements.txt
numpy
Pillow
matplotlib
fastapi
pydantic
uvicorn